transform registers, obj loading

+++ jni/core/vec3.h
@@ -23,8 +23,8 @@
 class vec3
 {
 public:
-    vec3() 
-    { 
+    vec3()
+    {
         x=y=z=0.0f;
 #ifdef _EE
         _ = 0.0f;
@@ -33,124 +33,124 @@ public:
 
     vec3(flx_real X, flx_real Y, flx_real Z) { x=X; y=Y; z=Z; }
     vec3(vec3 const &c) { *this=c; }
-		
+
     bool operator==(vec3 const &rhs) { return (x==rhs.x&&y==rhs.y&&z==rhs.z); }
-	
+
     inline vec3 &operator=(vec3 const &rhs)
     {
         x=rhs.x; y=rhs.y; z=rhs.z;
         return *this;
     }
-    
+
     inline vec3 operator+(vec3 const &rhs) const
     {
         vec3 t;
-        t.x=x+rhs.x; t.y=y+rhs.y; t.z=z+rhs.z; 
+        t.x=x+rhs.x; t.y=y+rhs.y; t.z=z+rhs.z;
         return t;
     }
-    
+
     inline vec3 operator-(vec3 const &rhs) const
     {
         vec3 t;
-        t.x=x-rhs.x; t.y=y-rhs.y; t.z=z-rhs.z; 
+        t.x=x-rhs.x; t.y=y-rhs.y; t.z=z-rhs.z;
         return t;
     }
-    
+
     inline vec3 operator*(vec3 const &rhs) const
     {
         vec3 t;
-        t.x=x*rhs.x; 
-        t.y=y*rhs.y; 
-        t.z=z*rhs.z; 
+        t.x=x*rhs.x;
+        t.y=y*rhs.y;
+        t.z=z*rhs.z;
         return t;
     }
-    
+
     inline vec3 operator/(vec3 const &rhs) const
     {
         vec3 t;
-        t.x=x/rhs.x; 
-        t.y=y/rhs.y; 
-        t.z=z/rhs.z; 
+        t.x=x/rhs.x;
+        t.y=y/rhs.y;
+        t.z=z/rhs.z;
         return t;
     }
-    
+
     inline vec3 operator+(flx_real rhs) const
     {
         vec3 t;
-        t.x=x+rhs; t.y=y+rhs; t.z=z+rhs; 
+        t.x=x+rhs; t.y=y+rhs; t.z=z+rhs;
         return t;
     }
-    
+
     inline vec3 operator-(flx_real rhs) const
     {
         vec3 t;
-        t.x=x-rhs; t.y=y-rhs; t.z=z-rhs; 
+        t.x=x-rhs; t.y=y-rhs; t.z=z-rhs;
         return t;
     }
-    
+
     inline vec3 operator*(flx_real rhs) const
     {
         vec3 t;
-        t.x=x*rhs; t.y=y*rhs; t.z=z*rhs; 
+        t.x=x*rhs; t.y=y*rhs; t.z=z*rhs;
         return t;
     }
-    
+
     inline vec3 operator/(flx_real rhs) const
     {
         vec3 t;
-        t.x=x/rhs; t.y=y/rhs; t.z=z/rhs; 
+        t.x=x/rhs; t.y=y/rhs; t.z=z/rhs;
         return t;
     }
-    
+
     inline vec3 &operator+=(vec3 const &rhs)
     {
-        x+=rhs.x; y+=rhs.y; z+=rhs.z; 
+        x+=rhs.x; y+=rhs.y; z+=rhs.z;
         return *this;
     }
-    
+
     inline vec3 &operator-=(vec3 const &rhs)
     {
-        x-=rhs.x; y-=rhs.y; z-=rhs.z; 
+        x-=rhs.x; y-=rhs.y; z-=rhs.z;
         return *this;
     }
-    
+
     inline vec3 &operator*=(flx_real rhs)
     {
-        x*=rhs; y*=rhs; z*=rhs; 
+        x*=rhs; y*=rhs; z*=rhs;
         return *this;
     }
-    
+
     inline vec3 &operator/=(flx_real rhs)
     {
         if (rhs!=0.0f) {x/=rhs; y/=rhs; z/=rhs;}
         return *this;
     }
-    
+
     inline flx_real dot(vec3 const &rhs) const
     {
         return x*rhs.x+y*rhs.y+z*rhs.z;
     }
-    
+
     inline vec3 cross(vec3 const &rhs) const
     {
         return vec3(y*rhs.z - z*rhs.y,
                     z*rhs.x - x*rhs.z,
                     x*rhs.y - y*rhs.x);
     }
-    
+
     inline vec3 reflect(vec3 const &rhs) const
     {
         flx_real vdn=dot(rhs)*2.0f;
         return (*this)-rhs*vdn;
     }
-    
+
     inline flx_real dist(vec3 const &rhs) const
     {
-        return sqrt((rhs.x-x)*(rhs.x-x)+
-                    (rhs.y-y)*(rhs.y-y)+
-                    (rhs.z-z)*(rhs.z-z));
+        return sqrt((float)((rhs.x-x)*(rhs.x-x)+
+                            (rhs.y-y)*(rhs.y-y)+
+                            (rhs.z-z)*(rhs.z-z)));
     }
-    
+
     inline flx_real distsq(vec3 const &rhs) const
     {
         return (rhs.x-x)*(rhs.x-x)+
@@ -163,7 +163,7 @@ public:
         // tendancy to blow up (overflow) in fixed...
         return sqrt((float)x*(float)x+(float)y*(float)y+(float)z*(float)z);
     }
-    
+
     inline flx_real magsq()
     {
         float xx=x;
@@ -171,22 +171,24 @@ public:
         float zz=z;
         return ((xx*xx)+(yy*yy)+(zz*zz));
     }
-	
+
     inline bool feq(const vec3 &other, flx_real epsilon=0.001)
     {
-        return (fabs(x-other.x)<epsilon && fabs(y-other.y)<epsilon && fabs(z-other.z)<epsilon);
+        return (fabs((float)(x-other.x))<epsilon &&
+                fabs((float)(y-other.y))<epsilon &&
+                fabs((float)(z-other.z))<epsilon);
     }
-	
+
     inline bool operator<(vec3 rhs)
     {
 			return x<rhs.x && y<rhs.y && z<rhs.z;
     }
-    
+
     inline bool operator>(vec3 rhs)
     {
         return x>rhs.x && y>rhs.y && z>rhs.z;
     }
-	
+
     inline vec3 &normalise() { *this/=mag(); return *this; }
 
     char *as_str()
